wip refactor trades
I think the old structure for trades, offers, transactions, etc. is
pretty good, in that the data is pretty well normalized.  But it made it
a pain to set things up, because everything relied on going through a
high-level Trade object.  By having each transaction be aware of the
trader/receiver, I think it will be much easier to test.